+++ rencfsMultiplatform/src/androidMain/preview/rs/xor/rencfs/krencfs/ui/screen/walkthrough/ConfirmationScreenPreview.kt
@@ -20,7 +20,6 @@ fun ConfirmationScreenPreview() {
             ),
             onBack = {},
             onViewDashboard = {},
-            onUnlockFolder = {},
             isDesktop = false,
         )
     }
