+++ rencfsMultiplatform/src/commonMain/kotlin/rs/xor/rencfs/krencfs/domain/walkthrough/PasswordStrength.kt
@@ -1,0 +1,7 @@
+package rs.xor.rencfs.krencfs.domain.walkthrough
+
+enum class PasswordStrength {
+    WEAK,
+    MEDIUM,
+    STRONG,
+}
