+++ rencfsMultiplatform/src/commonMain/sqldelight/rs/xor/rencfs/krencfs/1.sqm
@@ -1,0 +1,6 @@
+CREATE TABLE Vault (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    name TEXT NOT NULL,
+    path TEXT NOT NULL,
+    mount TEXT NOT NULL
+);
